Moved and renamed components to match other project items
